add functions to split text containing images or links into seperate nodes
