+++ src/main.py
@@ -1,14 +1,14 @@
 from textnode import TextNode
-from split_nodes import split_nodes_delimiter
+from inline_markdown import split_nodes_delimiter, split_nodes_image, split_nodes_links
 from htmlnode import HTMLNode
 
 def main():
-    test_node = TextNode("This is text with a `code block` word.", "code")
-    # htmlnode = test_node.text_node_to_html_node()
-    # print(htmlnode)
-    test_node = TextNode("This is text with a **bolded word** and **another**", "text")
-    result = split_nodes_delimiter([test_node], "**", "b")
-    # result = split_nodes_delimiter([test_node], "`","code")
+    test_node =  TextNode(
+        '''This is text with an [link 1](https://testing.com) and another [link 2](https://further-testing.com)''', "text"
+    )
+
+    result = split_nodes_links([test_node])
     print(result)
+
 
 main()
